requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.2
attrs==24.2.0
beautifulsoup4==4.12.2
Brotli==1.1.0
cachetools==5.5.0
CaptchaGenerator==1.1.6
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
cryptography==40.0.2
curl-cffi==0.5.10
decorator==4.4.2
dnspython==2.6.1
easygoogletranslate==0.0.4
emoji==2.12.1
fake-useragent==1.5.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.62.3
gTTS==2.5.3
h11==0.14.0
heroku3==5.2.1
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
lxml==4.9.4
moviepy==1.0.3
multidict==6.0.5
mutagen==1.47.0
numpy==2.1.0
oauthlib==3.2.2
packaging==23.1
patool==2.4.0
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==4.25.4
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.8.2
pydantic_core==2.20.1
pymediainfo-pyrofork==6.0.1
pymongo==4.8.0
pyparsing==3.1.2
pyrate-limiter==3.7.0
pyrofork
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
speedtest-cli==2.1.3
TgCrypto
tqdm==4.65.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==13.0
wget==3.2
yarl==1.9.4
yt-dlp
